+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.1.6
+
+- Improved: Show input indicator in the chat overview as well 
+- Improved: Username change error handling
+- Fix: Phantom push notification 
+- Fix: Start in chat, if configured
+- Fix: Smaller UI fixes
+
 ## 0.1.5
 
 - Fix: Reupload of media files was not working properly